dockerfiles/backend/repo/server/src/main/java/com/softserveinc/itacademy/bikechampionship/server/service/UserServiceImpl.java
package com.softserveinc.itacademy.bikechampionship.server.service;

import com.softserveinc.itacademy.bikechampionship.commons.exception.BadRequestException;
import com.softserveinc.itacademy.bikechampionship.commons.exception.ResourceNotFoundException;
import com.softserveinc.itacademy.bikechampionship.commons.model.User;
import com.softserveinc.itacademy.bikechampionship.commons.repository.UserRepository;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;


@Service
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;

    public UserServiceImpl(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public Boolean isEmailAvailable(String email) {
        return !userRepository.existsByEmail(email);
    }

    @Override
    public User getUserById(Long id) {
        return userRepository
            .findById(id)
            .orElseThrow(() -> new ResourceNotFoundException("User", "id", id));
    }

    @Override
    @Transactional
    public User editUserProfile(User user) {
        User userToUpdate = getUserById(user.getId());

        if (!user.getEmail().equals(userToUpdate.getEmail()) && isEmailAvailable(user.getEmail())) {
            userToUpdate.setEmail(user.getEmail());
            userToUpdate.setFirstName(user.getFirstName());
            userToUpdate.setLastName(user.getLastName());
            userToUpdate.setPhoneNumber(user.getPhoneNumber());
            return userRepository.save(userToUpdate);
        } else {
            throw new BadRequestException("This email is unavailable");
        }
    }

    @Override
    @Transactional
    public User changeUserPassword(User user) {
        User userToUpdate = getUserById(user.getId());
        userToUpdate.setPassword(user.getPassword());
        userRepository.save(userToUpdate);
        return userToUpdate;
    }

    @Override
    public void saveUser(User user) {
        userRepository.save(user);
    }
}
